removed dart:cli usage, use dylib version from pubspec.lock
